03-aidd/pyproject.toml
[project]
name = "aidd-bot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "aiogram==3.*",
  "openai>=1.44.0",
  "python-dotenv>=1.0.1",
]

[tool.uv]


